+++ src/backend/executor/nodeGroup.c
@@ -418,10 +418,10 @@ sameGroup(TupleTableSlot *oldslot,
 
 			val1 = fmgr(typoutput, attr1,
 						gettypelem(tupdesc->attrs[att - 1]->atttypid),
-						      (int)tupdesc->attrs[att - 1]->atttypmod);
+						      	   tupdesc->attrs[att - 1]->atttypmod);
 			val2 = fmgr(typoutput, attr2,
 						gettypelem(tupdesc->attrs[att - 1]->atttypid),
-							  (int)tupdesc->attrs[att - 1]->atttypmod);
+								   tupdesc->attrs[att - 1]->atttypmod);
 
 			/*
 			 * now, val1 and val2 are ascii representations so we can use
